requirements.txt
python-magic>=0.4.*, <=1
lxml>=4.0,<=5
